Bugfix: SpeedSensor must be of type Double.

+++ app/src/main/java/com/atrainingtracker/banalservice/devices/ant_plus/ANTBikeSpeedAndCadenceDevice.java
@@ -50,7 +50,7 @@ public class ANTBikeSpeedAndCadenceDevice extends MyANTDevice {
     protected AntPlusBikeSpeedDistancePcc bikeSpeedDistancePcc = null;
     protected AntPlusBikeCadencePcc cadencePcc = null;
     protected PccReleaseHandle pccReleaseHandle2 = null;
-    protected MySensor<BigDecimal> mSpeedSensor;
+    protected MySensor<Double> mSpeedSensor;
     protected MySensor<Double> mPaceSensor;
     protected MyDoubleAccumulatorSensor mDistanceSensor;
     protected MyDoubleAccumulatorSensor mLapDistanceSensor;
@@ -66,7 +66,7 @@ public class ANTBikeSpeedAndCadenceDevice extends MyANTDevice {
 
     @Override
     protected void addSensors() {
-        mSpeedSensor = new ThresholdSensor<BigDecimal>(this, SensorType.SPEED_mps, SPEED_THRESHOLD);
+        mSpeedSensor = new ThresholdSensor<Double>(this, SensorType.SPEED_mps, SPEED_THRESHOLD);
         mPaceSensor = new ThresholdSensor<Double>(this, SensorType.PACE_spm, SPEED_THRESHOLD);
         mDistanceSensor = new MyDoubleAccumulatorSensor(this, SensorType.DISTANCE_m, false);
         mLapDistanceSensor = new MyDoubleAccumulatorSensor(this, SensorType.DISTANCE_m_LAP, false);
@@ -129,7 +129,7 @@ public class ANTBikeSpeedAndCadenceDevice extends MyANTDevice {
 
                 @Override
                 public void onNewCalculatedSpeed(long estTimestamp, EnumSet<EventFlag> eventFlags, BigDecimal calculatedSpeed) {
-                    mSpeedSensor.newValue(calculatedSpeed);
+                    mSpeedSensor.newValue(calculatedSpeed.doubleValue());
                     mPaceSensor.newValue(1 / calculatedSpeed.doubleValue());   // TODO: correct?
                 }
             });
